+++ classes/completion/ollama.php
@@ -15,7 +15,7 @@
 // along with Moodle.  If not, see <http://www.gnu.org/licenses/>.
 
 /**
- * Version information
+ * Class providing completions for chat models (3.5 and up)
  *
  * @package    block_exaaichat
  * @copyright  2025 GTN Solutions https://gtn-solutions.com
@@ -23,10 +23,33 @@
  * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
 
-defined('MOODLE_INTERNAL') || die();
+namespace block_exaaichat\completion;
 
-$plugin->component = 'block_exaaichat';
-$plugin->version = 2025112800;
-$plugin->requires = 2022041600;
-$plugin->maturity = MATURITY_STABLE;
-$plugin->release = '3.2';
+defined('MOODLE_INTERNAL') || die;
+
+class ollama extends chat {
+    public function get_models(): array {
+        // liste von https://ollama.com/api/tags
+        $models = [
+            "cogito-2.1:671b",
+            "cogito-2.1:671b",
+            "glm-4.6",
+            "kimi-k2:1t",
+            "kimi-k2-thinking",
+            "qwen3-coder:480b",
+            "deepseek-v3.1:671b",
+            "gpt-oss:120b",
+            "gpt-oss:20b",
+            "qwen3-vl:235b-instruct",
+            "qwen3-vl:235b",
+            "minimax-m2",
+            "gemini-3-pro-preview",
+        ];
+
+        return array_combine($models, $models);
+    }
+
+    protected function get_default_endpoint(): string {
+        return 'https://ollama.com/v1/chat/completions';
+    }
+}